Refactored and renamed input and movement controllers
