+++ Runtime/Input/Keys/KeyDownInputController.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 
-namespace SkelTech.RPEST.Input.Keyboard {
-    public class KeyboardDownInputController : KeyboardInputController {
+namespace SkelTech.RPEST.Input.Keys {
+    public class KeyDownInputController : KeyInputController {
         #region Getters
         protected override bool IsInputKeyActive(KeyCode key) {
             return UnityEngine.Input.GetKeyDown(key);
