+++ Runtime/Input/Keys/KeyHoldInputController.cs
@@ -1,7 +1,7 @@
 using UnityEngine;
 
-namespace SkelTech.RPEST.Input.Keyboard {
-    public class KeyboardHoldInputController : KeyboardInputController {
+namespace SkelTech.RPEST.Input.Keys {
+    public class KeyHoldInputController : KeyInputController {
         #region Getters
         protected override bool IsInputKeyActive(KeyCode key) {
             return UnityEngine.Input.GetKey(key);
